UPSTREAM: arm64: mm: remove dsb from update_mmu_cache
update_mmu_cache() consists of a dsb(ishst) instruction so that new user
mappings are guaranteed to be visible to the page table walker on
exception return.

In reality this can be a very expensive operation which is rarely needed.
Removing this barrier shows a modest improvement in hackbench scores and
, in the worst case, we re-take the user fault and establish that there
was nothing to do.

Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>

Bug: 31432001
Change-Id: Ifb71e6b0154dacdd6ec380b897e962ffd0eb5ad9
(cherry picked from commit 120798d2e7d1ac87365fe5ea91b074bb42ca1eff)
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>